添加详情地图

+++ src/main/java/com/macro/mall/tiny/common/api/CommonResult.java
@@ -5,14 +5,14 @@ package com.macro.mall.tiny.common.api;
  * Created by macro on 2019/4/19.
  */
 public class CommonResult<T> {
-    private Integer code;
+    private int code;
     private String message;
     private T data;
 
     protected CommonResult() {
     }
 
-    protected CommonResult(Integer code, String message, T data) {
+    protected CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
@@ -98,7 +98,7 @@ public class CommonResult<T> {
         return new CommonResult<T>(ResultCode.FORBIDDEN.getCode(), ResultCode.FORBIDDEN.getMessage(), data);
     }
 
-    public long getCode() {
+    public int getCode() {
         return code;
     }
